Move ssh code to new package
