+++ command/ca/ca.go
@@ -84,7 +84,6 @@ $ step ca renew internal.crt internal.key \
 			rootComand(),
 			rootsCommand(),
 			federationCommand(),
-			sshCertificateCommand(),
 		},
 	}
